# Trade the ability to call db::insert during stack unwinding to fewer …